extract useSeed24Tabs
