+++ src/components/seed/24/index.ts
@@ -1,0 +1,1 @@
+export { default as useSeed24Tabs } from './useSeed24Tabs';
